refactor: extract help + search overlays (#357)
Eleventh slice. Two larger overlays bundled:

- draw_help    (~190 lines) - commands / shortcuts / keybindings / CLI options
- draw_search  (~150 lines) - message search results with snippet + match highlighting

Moved with the search overlay:
- search_snippet      (was pub(crate), now pub(in crate::ui))
- highlight_match_spans (private helper)
- search_snippet_*    tests (now in search.rs's tests module)

KeyAction is now only used by overlays/help.rs, so its import moved
out of ui/mod.rs.

ui/mod.rs: 3564 -> 3110 lines.
