+++ src/ui/mod.rs
@@ -18,12 +18,14 @@ use overlays::customize::draw_customize;
 use overlays::delete_confirm::draw_delete_confirm;
 use overlays::emoji_picker::draw_emoji_picker;
 use overlays::forward::draw_forward;
+use overlays::help::draw_help;
 use overlays::keybindings::draw_keybindings;
 use overlays::message_request::draw_message_request;
 use overlays::pin_duration::draw_pin_duration_picker;
 use overlays::poll_vote::draw_poll_vote_overlay;
 use overlays::profile::draw_profile;
 use overlays::reaction_picker::draw_reaction_picker;
+use overlays::search::draw_search;
 use overlays::settings::draw_settings;
 use overlays::settings_profile::draw_settings_profile_manager;
 use overlays::theme_picker::draw_theme_picker;
@@ -45,7 +47,6 @@ use ratatui::{
 use crate::app::{App, AutocompleteMode, GroupMenuState, InputMode, OverlayKind, VisibleImage};
 use crate::image_render::{self, ImageProtocol};
 use crate::input::{COMMANDS, format_compact_duration};
-use crate::keybindings::KeyAction;
 use crate::signal::types::{MessageStatus, PollData, PollVote, Reaction, StyleType, TrustLevel};
 use crate::theme::Theme;
 
@@ -61,8 +62,8 @@ pub(super) const CONTACTS_POPUP_WIDTH: u16 = 50;
 pub(super) const CONTACTS_MAX_VISIBLE: usize = 20;
 const FILE_BROWSER_POPUP_WIDTH: u16 = 60;
 const FILE_BROWSER_MAX_VISIBLE: usize = 20;
-const SEARCH_POPUP_WIDTH: u16 = 60;
-const SEARCH_MAX_VISIBLE: usize = 15;
+pub(super) const SEARCH_POPUP_WIDTH: u16 = 60;
+pub(super) const SEARCH_MAX_VISIBLE: usize = 15;
 const GROUP_MENU_POPUP_WIDTH: u16 = 40;
 const GROUP_MEMBER_MAX_VISIBLE: usize = 15;
 pub(super) const ABOUT_POPUP_WIDTH: u16 = 50;
@@ -2061,197 +2062,6 @@ fn draw_autocomplete(frame: &mut Frame, app: &App, input_area: Rect) {
     frame.render_widget(popup, area);
 }
 
-fn draw_help(frame: &mut Frame, app: &App, area: Rect) {
-    let theme = &app.theme;
-    let kb = &app.keybindings;
-
-    // Help table entries: (key, description)
-    let commands: &[(&str, &str)] = &[
-        ("/join <name>", "Switch to a conversation"),
-        ("/part", "Leave current conversation"),
-        ("/attach", "Attach a file"),
-        ("/search <query>", "Search messages"),
-        ("/sidebar", "Toggle sidebar visibility"),
-        ("/bell [type]", "Toggle notifications"),
-        ("/mute", "Mute/unmute conversation"),
-        ("/contacts", "Browse contacts"),
-        ("/settings", "Open settings"),
-        ("/keybindings", "Configure keybindings"),
-        ("/quit", "Exit siggy"),
-    ];
-
-    // Dynamic shortcuts from active keybindings
-    let dk = |a: KeyAction| kb.display_key(a);
-    let nav_keys = format!(
-        "{} / {}",
-        dk(KeyAction::NextConversation),
-        dk(KeyAction::PrevConversation)
-    );
-    let scroll_keys = format!(
-        "{} / {}",
-        dk(KeyAction::PageScrollUp),
-        dk(KeyAction::PageScrollDown)
-    );
-    let resize_keys = format!(
-        "{} / {}",
-        dk(KeyAction::ResizeSidebarLeft),
-        dk(KeyAction::ResizeSidebarRight)
-    );
-    let quit_key = dk(KeyAction::Quit);
-    let shortcuts: Vec<(String, &str)> = vec![
-        (nav_keys, "Next / prev conversation"),
-        ("Up / Down".to_string(), "Recall input history"),
-        ("@".to_string(), "Mention autocomplete"),
-        (scroll_keys, "Scroll messages"),
-        (resize_keys, "Resize sidebar"),
-        (quit_key, "Quit"),
-    ];
-
-    let cli: &[(&str, &str)] = &[
-        ("--incognito", "No local message storage"),
-        ("--demo", "Launch with dummy data"),
-        ("--setup", "Re-run first-time wizard"),
-    ];
-
-    // Dynamic normal-mode keybindings
-    let exit_key = dk(KeyAction::ExitInsert);
-    let insert_keys = format!(
-        "{} / {} / {} / {} / {}",
-        dk(KeyAction::InsertAtCursor),
-        dk(KeyAction::InsertAfterCursor),
-        dk(KeyAction::InsertLineStart),
-        dk(KeyAction::InsertLineEnd),
-        dk(KeyAction::OpenLineBelow)
-    );
-    let scroll_ud = format!(
-        "{} / {}",
-        dk(KeyAction::ScrollDown),
-        dk(KeyAction::ScrollUp)
-    );
-    let focus_ud = format!(
-        "{} / {}",
-        dk(KeyAction::FocusNextMessage),
-        dk(KeyAction::FocusPrevMessage)
-    );
-    let top_bottom = format!(
-        "{} / {}",
-        dk(KeyAction::ScrollToTop),
-        dk(KeyAction::ScrollToBottom)
-    );
-    let half_page = format!(
-        "{} / {}",
-        dk(KeyAction::HalfPageDown),
-        dk(KeyAction::HalfPageUp)
-    );
-    let cursor_lr = format!(
-        "{} / {}",
-        dk(KeyAction::CursorLeft),
-        dk(KeyAction::CursorRight)
-    );
-    let word_fb = format!(
-        "{} / {}",
-        dk(KeyAction::WordForward),
-        dk(KeyAction::WordBack)
-    );
-    let line_se = format!("{} / {}", dk(KeyAction::LineStart), dk(KeyAction::LineEnd));
-    let del_keys = format!(
-        "{} / {}",
-        dk(KeyAction::DeleteChar),
-        dk(KeyAction::DeleteToEnd)
-    );
-    let copy_keys = format!(
-        "{} / {}",
-        dk(KeyAction::CopyMessage),
-        dk(KeyAction::CopyAllMessages)
-    );
-    let search_keys = format!(
-        "{} / {}",
-        dk(KeyAction::NextSearchResult),
-        dk(KeyAction::PrevSearchResult)
-    );
-
-    let profile_label = format!("  Keybindings [{}]", app.keybindings.profile_name);
-    let vim: Vec<(String, &str)> = vec![
-        (exit_key, "Normal mode"),
-        (insert_keys, "Insert mode"),
-        (dk(KeyAction::InsertNewline), "Insert newline in input"),
-        (scroll_ud, "Scroll up / down"),
-        (focus_ud, "Prev / next message"),
-        (top_bottom, "Top / bottom of messages"),
-        (half_page, "Half-page scroll"),
-        (cursor_lr, "Cursor left / right"),
-        (word_fb, "Word forward / back"),
-        (line_se, "Start / end of line"),
-        (del_keys, "Delete char / to end"),
-        (copy_keys, "Copy message / full line"),
-        (dk(KeyAction::React), "React to focused message"),
-        (dk(KeyAction::Quote), "Reply / quote message"),
-        (dk(KeyAction::EditMessage), "Edit own message"),
-        (dk(KeyAction::DeleteMessage), "Delete message"),
-        (search_keys, "Next / prev search match"),
-        (dk(KeyAction::StartSearch), "Start command input"),
-    ];
-
-    // Calculate popup size
-    let key_col_width = 20;
-    let desc_col_width = 28;
-    let pref_width = (key_col_width + desc_col_width + 6) as u16;
-    let content_lines = commands.len() + shortcuts.len() + vim.len() + cli.len() + 7;
-    let pref_height = content_lines as u16 + 2;
-
-    let (popup_area, block) = centered_popup(frame, area, pref_width, pref_height, " Help ", theme);
-
-    let header_style = Style::default()
-        .fg(theme.accent_secondary)
-        .add_modifier(Modifier::BOLD);
-    let key_style = Style::default().fg(theme.accent);
-    let desc_style = Style::default().fg(theme.fg_secondary);
-
-    let mut lines: Vec<Line> = Vec::new();
-
-    let push_row = |lines: &mut Vec<Line>, key: &str, desc: &str| {
-        lines.push(Line::from(vec![
-            Span::styled(
-                format!("  {:<width$}", key, width = key_col_width),
-                key_style,
-            ),
-            Span::styled(desc.to_string(), desc_style),
-        ]));
-    };
-
-    lines.push(Line::from(Span::styled("  Commands", header_style)));
-    for &(key, desc) in commands {
-        push_row(&mut lines, key, desc);
-    }
-
-    lines.push(Line::from(""));
-    lines.push(Line::from(Span::styled("  Shortcuts", header_style)));
-    for (key, desc) in &shortcuts {
-        push_row(&mut lines, key, desc);
-    }
-
-    lines.push(Line::from(""));
-    lines.push(Line::from(Span::styled(&profile_label, header_style)));
-    for (key, desc) in &vim {
-        push_row(&mut lines, key, desc);
-    }
-
-    lines.push(Line::from(""));
-    lines.push(Line::from(Span::styled("  CLI Options", header_style)));
-    for &(key, desc) in cli {
-        push_row(&mut lines, key, desc);
-    }
-
-    lines.push(Line::from(""));
-    lines.push(Line::from(Span::styled(
-        "  Press any key to close",
-        Style::default().fg(theme.fg_muted),
-    )));
-
-    let popup = Paragraph::new(lines).block(block);
-    frame.render_widget(popup, popup_area);
-}
-
 fn draw_verify(frame: &mut Frame, app: &App, area: Rect) {
     let theme = &app.theme;
     let is_group = app
@@ -2453,250 +2263,9 @@ fn format_safety_number(sn: &str) -> Vec<String> {
     chunks.chunks(6).map(|row| row.join(" ")).collect()
 }
 
-fn draw_search(frame: &mut Frame, app: &App, area: Rect) {
-    let theme = &app.theme;
-    let max_visible = SEARCH_MAX_VISIBLE.min(app.search.results.len().max(1));
-    let pref_height = max_visible as u16 + 5; // +3 border/title +2 footer
-
-    let title = if app.search.query.is_empty() {
-        " Search ".to_string()
-    } else {
-        format!(" Search [{}] ", app.search.query)
-    };
-
-    let (popup_area, block) =
-        centered_popup(frame, area, SEARCH_POPUP_WIDTH, pref_height, &title, theme);
-
-    let inner_height = popup_area.height.saturating_sub(2) as usize; // minus borders
-    let footer_lines = 2; // footer + empty line
-    let visible_rows = inner_height.saturating_sub(footer_lines);
-
-    // Scroll the list so the selected item is always visible
-    let scroll_offset = if app.search.index >= visible_rows {
-        app.search.index - visible_rows + 1
-    } else {
-        0
-    };
-
-    let mut lines: Vec<Line> = Vec::new();
-    let inner_w = popup_area.width.saturating_sub(2) as usize;
-
-    if app.search.results.is_empty() {
-        let msg = if app.search.query.is_empty() {
-            "  Type to search..."
-        } else {
-            "  No results found"
-        };
-        lines.push(Line::from(Span::styled(
-            msg,
-            Style::default().fg(theme.fg_muted),
-        )));
-    } else {
-        let end = (scroll_offset + visible_rows).min(app.search.results.len());
-
-        for (i, result) in app.search.results[scroll_offset..end].iter().enumerate() {
-            let actual_index = scroll_offset + i;
-            let is_selected = actual_index == app.search.index;
-
-            // Format: [conv_name] sender: body_snippet
-            let conv_prefix = if app.active_conversation.is_some() {
-                String::new()
-            } else {
-                format!("[{}] ", truncate(&result.conv_name, 12))
-            };
-
-            let sender_display = truncate(&result.sender, 10);
-            let prefix = format!("  {conv_prefix}{sender_display}: ");
-            let body_max = inner_w.saturating_sub(prefix.len());
-            // Show a snippet of the body around the match
-            let body_snippet = search_snippet(&result.body, &app.search.query, body_max);
-
-            let prefix_style = if is_selected {
-                Style::default().bg(theme.bg_selected).fg(theme.accent)
-            } else {
-                Style::default().fg(theme.accent)
-            };
-            let body_style = if is_selected {
-                Style::default().bg(theme.bg_selected).fg(theme.fg)
-            } else {
-                Style::default().fg(theme.fg_secondary)
-            };
-
-            // Build spans with highlighted match
-            let mut spans = vec![Span::styled(prefix, prefix_style)];
-            spans.extend(highlight_match_spans(
-                &body_snippet,
-                &app.search.query,
-                body_style,
-                is_selected,
-                theme,
-            ));
-
-            lines.push(Line::from(spans));
-        }
-    }
-
-    // Pad to fill visible_rows so footer is always at the bottom
-    while lines.len() < visible_rows {
-        lines.push(Line::from(""));
-    }
-
-    lines.push(Line::from(""));
-    let count_text = if app.search.results.is_empty() {
-        String::new()
-    } else {
-        format!("  {}/{}", app.search.index + 1, app.search.results.len())
-    };
-    lines.push(Line::from(vec![
-        Span::styled(count_text, Style::default().fg(theme.warning)),
-        Span::styled(
-            "  j/k nav | Enter jump | n/N cycle | Esc close",
-            Style::default().fg(theme.fg_muted),
-        ),
-    ]));
-
-    let popup = Paragraph::new(lines).block(block);
-    frame.render_widget(popup, popup_area);
-}
-
 /// Extract a snippet of text centered around the first match of `query`.
-pub(crate) fn search_snippet(body: &str, query: &str, max_len: usize) -> String {
-    // Search results display on a single line; collapse newlines.
-    let body = body.replace('\n', " ");
-    let body = body.as_str();
-    let char_count = body.chars().count();
-    if char_count <= max_len {
-        return body.to_string();
-    }
-
-    // Find the match position in char indices (case-insensitive)
-    let body_lower = body.to_lowercase();
-    let query_lower = query.to_lowercase();
-    let match_byte_pos = body_lower.find(&query_lower).unwrap_or(0);
-    // Convert byte position in lowered text to char index
-    let match_char_pos = body_lower[..match_byte_pos].chars().count();
-
-    // Center the snippet around the match
-    let half = max_len / 2;
-    let start = match_char_pos.saturating_sub(half);
-    let end = (start + max_len).min(char_count);
-    let start = if end == char_count {
-        end.saturating_sub(max_len)
-    } else {
-        start
-    };
-
-    let snippet: String = body.chars().skip(start).take(end - start).collect();
-    let mut result = snippet;
-    if start > 0 {
-        result = format!("…{}", result.chars().skip(1).collect::<String>());
-    }
-    if end < char_count {
-        let trimmed: String = result
-            .chars()
-            .take(result.chars().count().saturating_sub(1))
-            .collect();
-        result = format!("{trimmed}…");
-    }
-    result
-}
-
 /// Build spans with the matching portions highlighted.
 /// Uses character-level case-insensitive matching to avoid byte-boundary issues.
-fn highlight_match_spans<'a>(
-    text: &str,
-    query: &str,
-    base_style: Style,
-    is_selected: bool,
-    theme: &Theme,
-) -> Vec<Span<'a>> {
-    if query.is_empty() {
-        return vec![Span::styled(text.to_string(), base_style)];
-    }
-
-    let match_style = if is_selected {
-        Style::default()
-            .bg(theme.bg_selected)
-            .fg(theme.warning)
-            .add_modifier(Modifier::BOLD)
-    } else {
-        Style::default()
-            .fg(theme.warning)
-            .add_modifier(Modifier::BOLD)
-    };
-
-    // Find all match positions using the lowercased strings
-    let text_lower = text.to_lowercase();
-    let query_lower = query.to_lowercase();
-    let query_len = query_lower.len();
-
-    // Collect match byte ranges in the lowered text, then map back to original
-    // For ASCII and most Unicode, to_lowercase preserves byte length per-character.
-    // Use the lowered text offsets directly since they share the same structure.
-    let mut match_ranges: Vec<(usize, usize)> = Vec::new();
-    let mut search_pos = 0;
-    while search_pos < text_lower.len() {
-        if let Some(m) = text_lower[search_pos..].find(&query_lower) {
-            let start = search_pos + m;
-            let end = start + query_len;
-            match_ranges.push((start, end));
-            search_pos = end;
-        } else {
-            break;
-        }
-    }
-
-    if match_ranges.is_empty() {
-        return vec![Span::styled(text.to_string(), base_style)];
-    }
-
-    // Build a char-index mapping: for each char, record its byte start in original and lowered
-    let orig_chars: Vec<(usize, char)> = text.char_indices().collect();
-    let lower_chars: Vec<(usize, char)> = text_lower.char_indices().collect();
-
-    // Build byte-position mapping from lowered → original using char alignment
-    // Both should have the same number of chars
-    let char_count = orig_chars.len().min(lower_chars.len());
-
-    // Convert match_ranges from lowered byte positions to original byte positions
-    let mut orig_ranges: Vec<(usize, usize)> = Vec::new();
-    for &(low_start, low_end) in &match_ranges {
-        let start_char = lower_chars.iter().position(|&(pos, _)| pos == low_start);
-        let end_char = lower_chars
-            .iter()
-            .position(|&(pos, _)| pos == low_end)
-            .unwrap_or(char_count);
-        if let Some(sc) = start_char {
-            let orig_start = orig_chars[sc].0;
-            let orig_end = if end_char < orig_chars.len() {
-                orig_chars[end_char].0
-            } else {
-                text.len()
-            };
-            orig_ranges.push((orig_start, orig_end));
-        }
-    }
-
-    // Build spans from original ranges
-    let mut spans = Vec::new();
-    let mut pos = 0;
-    for (start, end) in orig_ranges {
-        if start > pos {
-            spans.push(Span::styled(text[pos..start].to_string(), base_style));
-        }
-        spans.push(Span::styled(text[start..end].to_string(), match_style));
-        pos = end;
-    }
-    if pos < text.len() {
-        spans.push(Span::styled(text[pos..].to_string(), base_style));
-    }
-
-    if spans.is_empty() {
-        spans.push(Span::styled(text.to_string(), base_style));
-    }
-    spans
-}
-
 /// Format a file size in human-readable form (B, K, M, G).
 pub(crate) fn format_file_size(bytes: u64) -> String {
     if bytes < 1024 {
@@ -3203,29 +2772,6 @@ mod tests {
     #[case(1_073_741_824, "1.0G")]
     fn format_file_size_cases(#[case] bytes: u64, #[case] expected: &str) {
         assert_eq!(format_file_size(bytes), expected);
-    }
-
-    // --- search_snippet ---
-
-    #[test]
-    fn search_snippet_short_passthrough() {
-        let body = "short text";
-        assert_eq!(search_snippet(body, "short", 100), body);
-    }
-
-    #[test]
-    fn search_snippet_centers_on_match() {
-        let body = "a".repeat(100) + "NEEDLE" + &"b".repeat(100);
-        let snippet = search_snippet(&body, "NEEDLE", 30);
-        assert!(
-            snippet.chars().count() <= 30,
-            "snippet too long ({} chars): {snippet}",
-            snippet.chars().count()
-        );
-        assert!(
-            snippet.contains("NEEDLE"),
-            "expected query in snippet: {snippet}"
-        );
     }
 }
